+++ server/src/api/routers/sessions/workout.py
@@ -1,5 +1,6 @@
 from fastapi import APIRouter, Depends
 from schemas.workout import WorkoutRequest, WorkoutResponse
+from schemas.workout import Created_Workout_Response
 from services.workout import WorkoutService
 from depends.auth import FirebaseToken
 
@@ -13,4 +14,17 @@ async def post_workout(workout_request: WorkoutRequest, fbToken: FirebaseToken =
     response: WorkoutResponse = {
         "result": message
     }
+    return response
+
+
+@router.get("/", response_model=Created_Workout_Response)
+async def get_workout_history_based_date(date: str,
+                                         fbToken: FirebaseToken = Depends()):
+    uid = fbToken.uid
+    date = date
+    response: Created_Workout_Response = {
+        "date": date,
+        "datas": await WorkoutService.get_based_on_date(uid, date + '%')
+    }
+    # response: TodayRequest = TodayService.get_today_record(uid, date)
     return response
