--- server/src/api/services/today.py
@@ -1,7 +1,0 @@
-import usecases.today
-
-
-class TodayService:
-    @staticmethod
-    async def get_today_record(uid, date):
-        return usecases.today.get_today_record(uid, date)
